requirements.txt
Flask==3.0.0
flask-cors==4.0.0
ccxt==4.4.87
pandas>=2.0.0
numpy>=1.25.0
plotly==5.17.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4 